requirements.txt
fastapi==0.101.0
jira==3.5.2
Jinja2==3.1.2
starlette==0.27.0
python-dotenv==1.0.0
